scenix_api/requirements.txt
Flask
mysql-connector-python
flask-cors
fastapi
uvicorn